+++ module-api/src/main/java/kernel/jdon/global/exception/ApiException.java
@@ -1,11 +1,11 @@
-package kernel.jdon.error.exception.api;
+package kernel.jdon.global.exception;
 
-import kernel.jdon.error.code.ErrorCode;
+import kernel.jdon.error.ErrorCode;
 import lombok.Getter;
 
 @Getter
 public class ApiException extends RuntimeException {
-	private transient ErrorCode errorCode;
+	private final transient ErrorCode errorCode;
 
 	public ApiException(ErrorCode errorCode) {
 		this.errorCode = errorCode;
