+++ module-crawler/src/main/java/kernel/jdon/crawler/config/AppConfig.java
@@ -1,4 +1,4 @@
-package kernel.jdon.config;
+package kernel.jdon.crawler.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
