+++ module-crawler/src/main/java/kernel/jdon/crawler/wanted/service/WantedCrawlerService.java
@@ -13,7 +13,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
-import kernel.jdon.config.UrlConfig;
+import kernel.jdon.crawler.config.UrlConfig;
+import kernel.jdon.crawler.global.error.code.WantedErrorCode;
+import kernel.jdon.crawler.global.error.exception.CrawlerException;
 import kernel.jdon.crawler.wanted.converter.EntityConverter;
 import kernel.jdon.crawler.wanted.dto.object.CreateSkillDto;
 import kernel.jdon.crawler.wanted.dto.response.WantedJobDetailResponse;
@@ -23,8 +25,6 @@ import kernel.jdon.crawler.wanted.search.JobSearchJobCategory;
 import kernel.jdon.crawler.wanted.search.JobSearchJobPosition;
 import kernel.jdon.crawler.wanted.search.JobSearchLocation;
 import kernel.jdon.crawler.wanted.search.JobSearchSort;
-import kernel.jdon.error.code.crawler.WantedErrorCode;
-import kernel.jdon.error.exception.crawler.CrawlerException;
 import kernel.jdon.jobcategory.domain.JobCategory;
 import kernel.jdon.jobcategory.repository.JobCategoryRepository;
 import kernel.jdon.skill.domain.Skill;
